requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
commonX==0.6.37
curl_cffi==0.10.0
Deprecated==1.2.18
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
img2pdf==0.6.0
jiter==0.9.0
jmcomic==2.5.32
linkify-it-py==2.0.3
loguru==0.7.3
lxml==5.3.1
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.2.0
nonebot-adapter-onebot==2.4.6
nonebot-plugin-apscheduler==0.5.0
nonebot2==2.4.2
nonechat==0.3.0
openai==1.73.0
packaging==24.2
pikepdf==9.5.2
pillow==11.1.0
platformdirs==4.3.7
propcache==0.3.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pygtrie==2.5.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
sniffio==1.3.1
starlette==0.46.1
textual==2.1.2
tqdm==4.67.1
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
win32_setctime==1.2.0
wrapt==1.17.2
yarl==1.18.3
pytz==2025.2
qrcode==8.2